test: snapshot of the transition from CircuitPython to TinyGo

+++ devices/raspberry_pi_pico_2w/tinygo/internal/usbcdc/enums/incoming_category.go
@@ -6,8 +6,8 @@ import (
 )
 
 type (
-	// IncomingCategory represent the enum categories of incoming messages from the Raspberry Pi 5
-	IncomingCategory int
+	// IncomingCategory represents the enum categories of incoming messages from the Raspberry Pi 5
+	IncomingCategory uint8
 )
 
 const (
